+++ routes/statusRoutes.js
@@ -1,7 +1,8 @@
 import express from "express";
-import { getStatus } from "../controllers/statusController.js";
+import { getStatus, getSysInfo } from "../controllers/statusController.js";
 const router = express.Router();
 
 router.get("/status", getStatus);
+router.get("/sysinfo", getSysInfo);
 
 export default router;
